# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Battleships API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://rudko-laravel/api",
            "description": "Laravel API Swagger"
        }
    ],
    "paths": {
        "/chat-load/{id}/{code}": {
            "get": {
                "tags": [
                    "API"
                ],
                "summary": "Loads a messages from database",
                "operationId": "Load messages",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game id",
                        "required": true
                    },
                    {
                        "name": "code",
                        "in": "path",
                        "description": "The user's code",
                        "required": true
                    },
                    {
                        "name": "lastTime",
                        "in": "query",
                        "description": "Timestamp of the most recent message",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat loaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "my": {
                                                        "type": "boolean",
                                                        "example": "true"
                                                    },
                                                    "time": {
                                                        "type": "integer",
                                                        "example": "1495517526"
                                                    },
                                                    "message": {
                                                        "type": "string",
                                                        "example": "Hello, World!"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "lastTime": {
                                            "type": "integer",
                                            "example": "61122735b150c"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/chat-send/{id}/{code}": {
            "post": {
                "tags": [
                    "API"
                ],
                "summary": "Sends a message to a user",
                "operationId": "Send message",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game id",
                        "required": true
                    },
                    {
                        "name": "code",
                        "in": "path",
                        "description": "The user's code",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "description": "Message to send (250 characters limit)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/clear-field/{id}/{code}": {
            "post": {
                "tags": [
                    "API"
                ],
                "summary": "Removes all ships from the field",
                "operationId": "Clear field",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game id",
                        "required": true
                    },
                    {
                        "name": "code",
                        "in": "path",
                        "description": "The user's code",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully cleared",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/place-ship/{id}/{code}": {
            "post": {
                "tags": [
                    "API"
                ],
                "summary": "Places/removes/turns/replaces ship",
                "operationId": "Place ship",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game id",
                        "required": true
                    },
                    {
                        "name": "code",
                        "in": "path",
                        "description": "The user's code",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "x": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "y": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "ship": {
                                        "type": "strind",
                                        "example": "1-1"
                                    },
                                    "orientation": {
                                        "type": "string",
                                        "example": "vertical"
                                    },
                                    "ships": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "x": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "y": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "ship": {
                                                    "type": "string",
                                                    "example": "1-1"
                                                },
                                                "orientation": {
                                                    "type": "string",
                                                    "example": "vertical"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/shot/{id}/{code}": {
            "post": {
                "tags": [
                    "API"
                ],
                "summary": "Shoots",
                "operationId": "Shots",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game id",
                        "required": true
                    },
                    {
                        "name": "code",
                        "in": "path",
                        "description": "The user's code",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "x",
                                    "y"
                                ],
                                "properties": {
                                    "x": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "y": {
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/start": {
            "post": {
                "tags": [
                    "API"
                ],
                "summary": "Creates a new game",
                "operationId": "Create a new game",
                "responses": {
                    "200": {
                        "description": "New game created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "61122735b150c"
                                        },
                                        "invited": {
                                            "type": "string",
                                            "example": "6112273ab92c1"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/status/{id}/{code}": {
            "post": {
                "tags": [
                    "API"
                ],
                "summary": "Retrieves game info",
                "operationId": "Retrieve status",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game id",
                        "required": true
                    },
                    {
                        "name": "code",
                        "in": "path",
                        "description": "The user's code",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat loaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "game": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "9"
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "invite": {
                                                    "type": "string",
                                                    "example": "aeGt81HUEn"
                                                },
                                                "myTurn": {
                                                    "type": "boolean",
                                                    "example": "true"
                                                },
                                                "meReady": {
                                                    "type": "boolean",
                                                    "example": "true"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "fieldMy": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "fieldEnemy": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "usedPlaces": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ready/{id}/{code}": {
            "post": {
                "tags": [
                    "API"
                ],
                "summary": "Get user ready",
                "operationId": "User ready",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game id",
                        "required": true
                    },
                    {
                        "name": "code",
                        "in": "path",
                        "description": "The user's code",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat loaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "API",
            "description": "Api methods"
        }
    ],
    "security": [
        []
    ]
}